+++ yee-kosari/src/App.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import "./App.css";
+import Router from "./components/Router";
+
+function App() {
+  return <Router />;
+}
+
+export default App;
